Add a Questions Page
This create the Questions component that allows
a user to actually vote on a question.  As part of this
we introduce the relevant action creators and reducers
to allow us to save answers to the store.  The shared
`handleUserVote` sends out the dispatch to both the
questions and users reducers so that they update.