Add action/reducer that adds new questions to the user state
The previous implementation had a bug where the question
would be added to the questions state, but not the authors
questions array.  So we fix this by adding a new ADD_QUESTION_TO_USER
action and a handleAddQuestion function that can dispatch
the changes to both the questions state and the users state.