Add a create profile page
Renders under the route "users/new".

This adds action creators and reducers to add a new
user.  It will navigate when you submit your new
profile to the page.  We also created a PasswordInput
component that can show/hide your password while you
input it.  I've reused it on the login page.

Additionally, I've made a couple things into
`box-sizing: border-box; `
which helps all the sizing on the pages.  Not
totally up to date on css practices, but hey,
I'll take it.