Enable saving redux state to the local store
First, we moved the store into it's own file.  This
lets us use a function to serialize state and store it
in the local storage.  Then it let's use use a function
to load from local storage.

Second, we update the Login component to remember
the location that was prior.  This way we can redirect back
to the previous location after logging in!