Update Question to better handle voting
When you vote on a quesiton, we now outline your
choice with a green box. This is a cleaner UI. Meanwhile,
we removed the UserAnswer box because it's no longer needed.

Then, for each answered question, show the percentage, the
number of votes, and the total vote count.

Also removed some uncessary CSS classes.

Fixes:

9. Upon voting in a poll, all of the information of the answered poll is
   displayed.
10. Question Answered: Please show the number of voters, along with the
    percentage. The number of people who voted for that option.

Test: `npm test`