Enable table layout of questions
This creates a tab layout format for questions.  We introduce a new component
QuestionsTabLayout that takes a set of tabs and questionIds to display in
each tab.  We can use this on the Dashboard component and the UserProfile
component.

Also update Question and CreateUser to use our lighter-background styling.
This makes the app feel a bit more uniform.

Finally, update snapshots for everything.

Fixes:

4. The user can alternate between viewing answered and unanswered polls.

- Both types of questions are displayed on the main page! Please add a button or
  tabs or something, with a state value, to allow the user to alternate between
  them, and make the "unanswered" polls displayed by default.

5. The unanswered questions are shown by default.