Passwords are no longer sent to the client
This ensures passwords cannot be collected at the client side.

In the chance, we also formalize the concept of a impersonable
user, which a pre-set list of users that you can impersonate to
check out functionality.  This is enforced by the backend so you
can't login as another user.  Someone could still technically
maliciously modify the redux store, which is where JWTs would
come into play.

Also, we ensure new passwords adhere to at least 6 characters.